NEXT-13468 - Remove app store php components

See merge request shopware/6/product/platform!4656

+++ src/Administration/Resources/app/administration/src/module/sw-login/view/sw-login-login/index.js
@@ -51,7 +51,9 @@ Component.register('sw-login-login', {
                 setTimeout(resolve, 150);
             });
 
-            this.licenseViolationService.removeTimeFromLocalStorage(this.licenseViolationService.key.showViolationsKey);
+            if (this.licenseViolationService) {
+                this.licenseViolationService.removeTimeFromLocalStorage(this.licenseViolationService.key.showViolationsKey);
+            }
 
             return animationPromise.then(() => {
                 this.$parent.isLoginSuccess = false;
